# [Perf: Init Costs] Batch initial store updates together + more