… packages/react/src/types/store.ts