@moklick 2023-12-17
As discussed, we only took a small part of this PR. The idea how we could re-write the `StoreUpdater` component was great and helped to make it easier to understand. Maintaining the `batchUpdate` middleware seems to costly for to little improvement here but it's a very good idea. Maybe we can add it in the future.  

closed by #3729